Resolved #57
